refactor(category): #1 CategoryIcon・CategoryColor 型を components/category/types.ts に移動
将来的に lib/types.ts が削除される際に UI 固有の型が残るよう、
アイコン・色識別子の型定義を components/category/types.ts に移動。
lib/types.ts からは re-export し既存の import パスを維持。
