+++ packages/app/frontend/components/category/types.ts
@@ -1,0 +1,35 @@
+export type CategoryIcon =
+  | 'tag'
+  | 'wallet'
+  | 'trending_up'
+  | 'trending_down'
+  | 'piggy_bank'
+  | 'house'
+  | 'utensils'
+  | 'shopping_cart'
+  | 'car'
+  | 'bus'
+  | 'plane'
+  | 'heart_pulse'
+  | 'graduation_cap'
+  | 'briefcase'
+  | 'music'
+  | 'zap'
+  | 'wifi'
+  | 'shirt'
+  | 'dumbbell'
+  | 'coffee'
+  | 'gift'
+  | 'book'
+  | 'baby'
+  | 'plus'
+
+export type CategoryColor =
+  | 'red'
+  | 'orange'
+  | 'yellow'
+  | 'green'
+  | 'teal'
+  | 'blue'
+  | 'purple'
+  | 'pink'
